refactor(feature:account): Move file to commonMain and resolve Detekt issues
