+++ feature/account/src/commonMain/composeResources/drawable/feature_account_error_black.xml
